# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jadi.ia API - Plataforma de criação de sites e aplicações com IA
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Project {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  language: string;
  theme: string;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
  language: string;
  theme?: string;
}

export interface UpdateProjectBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  language?: string;
  theme?: string;
  isPublic?: boolean;
}

export interface ProjectFile {
  id: number;
  projectId: number;
  name: string;
  content: string;
  language: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateFileBody {
  name: string;
  content?: string;
  language: string;
}

export interface UpdateFileBody {
  name?: string;
  content?: string;
  language?: string;
}

export interface ProjectSecret {
  id: number;
  projectId: number;
  key: string;
  createdAt: string;
}

export interface CreateSecretBody {
  key: string;
  value: string;
}

export interface AiMessage {
  role: string;
  content: string;
}

export interface AiChatBody {
  messages: AiMessage[];
  /** @nullable */
  projectContext?: string | null;
  /** @nullable */
  language?: string | null;
}

export interface AiChatResponse {
  message: string;
  codeBlocks: string[];
}

export interface GeneratePromptBody {
  description: string;
  projectType: string;
  language?: string;
}

export interface GeneratePromptResponse {
  prompt: string;
  suggestions: string[];
}

export type DashboardStatsLanguagesUsedItem = {
  language: string;
  count: number;
};

export interface DashboardStats {
  totalProjects: number;
  totalFiles: number;
  recentProjectsCount: number;
  languagesUsed: DashboardStatsLanguagesUsedItem[];
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  createdAt: string;
}