# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jadi.ia API - Plataforma de criação de sites e aplicações com IA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Registrar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email já utilizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login bem sucedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Dados do usuário autenticado
      responses:
        "200":
          description: Dados do usuário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # PROJECTS
...
  /projects/{id}/files:
    get:
      operationId: listProjectFiles
      tags: [files]
      summary: Listar arquivos de um projeto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de arquivos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"

    post:
      operationId: createProjectFile
      tags: [files]
      summary: Criar arquivo no projeto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileBody"
      responses:
        "201":
          description: Arquivo criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
  /projects/{id}/files/{fileId}:
    get:
      operationId: getProjectFile
      tags: [files]
      summary: Buscar arquivo específico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Arquivo encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
        "404":
          description: Arquivo não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateProjectFile
      tags: [files]
      summary: Atualizar conteúdo do arquivo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFileBody"
      responses:
        "200":
          description: Arquivo atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"

    delete:
      operationId: deleteProjectFile
      tags: [files]
      summary: Deletar arquivo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Arquivo deletado

  # SECRETS
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Enviar mensagem para IA (Groq)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: Resposta da IA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
        "500":
          description: Erro na IA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/generate-prompt:
    post:
      operationId: generatePrompt
      tags: [ai]
      summary: Agente técnico gera prompt otimizado para desenvolvimento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePromptBody"
      responses:
        "200":
          description: Prompt gerado pelo agente técnico
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratePromptResponse"

  # STATS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - createdAt

    # PROJECTS
...
    ProjectFile:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        content:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - content
        - language
        - createdAt
        - updatedAt
    CreateFileBody:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        language:
          type: string
      required:
        - name
        - language
    UpdateFileBody:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        language:
          type: string

    # SECRETS
...
    AiMessage:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
      required:
        - role
        - content
    AiChatBody:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AiMessage"
        projectContext:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
      required:
        - messages
    AiChatResponse:
      type: object
      properties:
        message:
          type: string
        codeBlocks:
          type: array
          items:
            type: string
      required:
        - message
        - codeBlocks
    GeneratePromptBody:
      type: object
      properties:
        description:
          type: string
        projectType:
          type: string
        language:
          type: string
      required:
        - description
        - projectType
    GeneratePromptResponse:
      type: object
      properties:
        prompt:
          type: string
        suggestions:
          type: array
          items:
            type: string
      required:
        - prompt
        - suggestions

    # STATS
...