# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jadi.ia API - Plataforma de criação de sites e aplicações com IA
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Registrar novo usuário
 */
export const RegisterBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
});

/**
 * @summary Login do usuário
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    createdAt: zod.coerce.date(),
  }),
  token: zod.string(),
});

/**
 * @summary Logout do usuário
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Dados do usuário autenticado
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Listar projetos do usuário
 */
export const ListProjectsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  language: zod.string(),
  theme: zod.string(),
  isPublic: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Criar novo projeto
 */
export const CreateProjectBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
  language: zod.string(),
  theme: zod.string().optional(),
});

/**
 * @summary Buscar projeto por ID
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  language: zod.string(),
  theme: zod.string(),
  isPublic: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Atualizar projeto
 */
export const UpdateProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProjectBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
  language: zod.string().optional(),
  theme: zod.string().optional(),
  isPublic: zod.boolean().optional(),
});

export const UpdateProjectResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  language: zod.string(),
  theme: zod.string(),
  isPublic: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Deletar projeto
 */
export const DeleteProjectParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Listar arquivos de um projeto
 */
export const ListProjectFilesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListProjectFilesResponseItem = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  name: zod.string(),
  content: zod.string(),
  language: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListProjectFilesResponse = zod.array(ListProjectFilesResponseItem);

/**
 * @summary Criar arquivo no projeto
 */
export const CreateProjectFileParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateProjectFileBody = zod.object({
  name: zod.string(),
  content: zod.string().optional(),
  language: zod.string(),
});

/**
 * @summary Buscar arquivo específico
 */
export const GetProjectFileParams = zod.object({
  id: zod.coerce.number(),
  fileId: zod.coerce.number(),
});

export const GetProjectFileResponse = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  name: zod.string(),
  content: zod.string(),
  language: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Atualizar conteúdo do arquivo
 */
export const UpdateProjectFileParams = zod.object({
  id: zod.coerce.number(),
  fileId: zod.coerce.number(),
});

export const UpdateProjectFileBody = zod.object({
  name: zod.string().optional(),
  content: zod.string().optional(),
  language: zod.string().optional(),
});

export const UpdateProjectFileResponse = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  name: zod.string(),
  content: zod.string(),
  language: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Deletar arquivo
 */
export const DeleteProjectFileParams = zod.object({
  id: zod.coerce.number(),
  fileId: zod.coerce.number(),
});

/**
 * @summary Listar secrets do projeto (apenas chaves)
 */
export const ListProjectSecretsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListProjectSecretsResponseItem = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  key: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListProjectSecretsResponse = zod.array(
  ListProjectSecretsResponseItem,
);

/**
 * @summary Criar secret no projeto
 */
export const CreateProjectSecretParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateProjectSecretBody = zod.object({
  key: zod.string(),
  value: zod.string(),
});

/**
 * @summary Deletar secret do projeto
 */
export const DeleteProjectSecretParams = zod.object({
  id: zod.coerce.number(),
  secretId: zod.coerce.number(),
});

/**
 * @summary Enviar mensagem para IA (Groq)
 */
export const AiChatBody = zod.object({
  messages: zod.array(
    zod.object({
      role: zod.string(),
      content: zod.string(),
    }),
  ),
  projectContext: zod.string().nullish(),
  language: zod.string().nullish(),
});

export const AiChatResponse = zod.object({
  message: zod.string(),
  codeBlocks: zod.array(zod.string()),
});

/**
 * @summary Agente técnico gera prompt otimizado para desenvolvimento
 */
export const GeneratePromptBody = zod.object({
  description: zod.string(),
  projectType: zod.string(),
  language: zod.string().optional(),
});

export const GeneratePromptResponse = zod.object({
  prompt: zod.string(),
  suggestions: zod.array(zod.string()),
});

/**
 * @summary Estatísticas do dashboard do usuário
 */
export const GetDashboardStatsResponse = zod.object({
  totalProjects: zod.number(),
  totalFiles: zod.number(),
  recentProjectsCount: zod.number(),
  languagesUsed: zod.array(
    zod.object({
      language: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Atividade recente do usuário
 */
export const GetRecentActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.string(),
  description: zod.string(),
  projectId: zod.number().nullish(),
  projectName: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);