# lib/api-zod/src/generated/types/generatePromptResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jadi.ia API - Plataforma de criação de sites e aplicações com IA
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratePromptResponse {
  prompt: string;
  suggestions: string[];
}